+++ tests/unit/test_experiments/test_simulation_with_experiment.py
@@ -1,4 +1,5 @@
 import copy
+import itertools
 import logging
 import os
 from datetime import datetime
@@ -864,35 +865,53 @@ class TestSimulationExperiment:
         experiment_2step = pybamm.Experiment(
             [
                 (
-                    "Discharge at C/20 for 1 hour",
-                    "Charge at 1 A until 4.1 V",
-                    "Hold at 4.1 V until C/2",
-                    "Discharge at 2 W for 1 hour",
+                    "Discharge at C/10 for 1 hour",
+                    "Charge at C/20 for 1 hour",
+                    "Hold at 3.7 V for 1 hour",
                 ),
             ]
             * 2,
+            period="100 hours",  # only capture the first and final values
         )
+        rtol = 1e-6
+        atol = 1e-15
 
-        solutions = []
-        for solver in [pybamm.CasadiSolver(), pybamm.IDAKLUSolver()]:
+        solvers = {
+            "casadi": pybamm.CasadiSolver(atol=atol, rtol=rtol),
+            "idaklu": pybamm.IDAKLUSolver(atol=atol, rtol=rtol),
+        }
+
+        solutions = {}
+        for name, solver in solvers.items():
             model = pybamm.lithium_ion.SPM()
             sim = pybamm.Simulation(model, experiment=experiment_2step, solver=solver)
             solution = sim.solve()
-            solutions.append(solution)
+            assert solution.t[-1] == pytest.approx(3600 * len(experiment_2step.steps))
+            solutions[name] = solution
 
-        np.testing.assert_allclose(
-            solutions[0]["Voltage [V]"].data,
-            solutions[1]["Voltage [V]"](solutions[0].t),
-            rtol=1e-2,
-            atol=1e-1,
-        )
-        np.testing.assert_allclose(
-            solutions[0]["Current [A]"].data,
-            solutions[1]["Current [A]"](solutions[0].t),
-            rtol=1e-0,
-            atol=1e-0,
-        )
-        assert solutions[1].termination == "final time"
+        num_sub_solutions = len(solutions["casadi"].sub_solutions)
+        assert len(solutions["idaklu"].sub_solutions) == num_sub_solutions
+
+        def get_sub_solution(name, idx_sol, idx_time):
+            y = solutions[name].sub_solutions[idx_sol].all_ys[0][:, idx_time]
+            if isinstance(y, casadi.DM):
+                y = y.full().flatten()
+            return y
+
+        for idx_sol, idx_time in itertools.product(range(num_sub_solutions), [0, -1]):
+            t = solutions["casadi"].sub_solutions[idx_sol].t[idx_time]
+            casadi_sol = get_sub_solution("casadi", idx_sol, idx_time)
+            idaklu_sol = get_sub_solution("idaklu", idx_sol, idx_time)
+
+            err_msg = f"Failed to match solution y values at {t} seconds"
+            np.testing.assert_allclose(
+                casadi_sol,
+                idaklu_sol,
+                rtol=rtol * 100,
+                atol=atol * 100,
+                err_msg=err_msg,
+            )
+        assert solutions["casadi"].termination == "final time"
 
     @pytest.mark.parametrize(
         "solver_cls",